Update per CR feedback.
