+++ assets/js/modules/sign-in-with-google/components/setup/SetupForm.js
@@ -19,6 +19,7 @@
 /**
  * WordPress dependencies
  */
+import { lazy, Suspense, createInterpolateElement } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
 
 /**
@@ -27,10 +28,16 @@ import { __ } from '@wordpress/i18n';
 import { useSelect } from 'googlesitekit-data';
 import StoreErrorNotices from '../../../../components/StoreErrorNotices';
 import { MODULES_SIGN_IN_WITH_GOOGLE } from '../../datastore/constants';
+import { CORE_SITE } from '../../../../googlesitekit/datastore/site/constants';
 import ClientIDTextField from '../common/ClientIDTextField';
 import Button from '../../../../googlesitekit/components-gm2/Button';
-import { CORE_SITE } from '../../../../googlesitekit/datastore/site/constants';
+import SupportLink from '../../../../components/SupportLink';
 import ExternalIcon from '../../../../../svg/icons/external.svg';
+import PreviewBlock from '../../../../components/PreviewBlock';
+import MediaErrorHandler from '../../../../components/MediaErrorHandler';
+const LazyGraphicSVG = lazy( () =>
+	import( '../../../../../svg/graphics/sign-in-with-google-setup.svg' )
+);
 
 export default function SetupForm() {
 	const siteName = useSelect( ( select ) =>
@@ -39,31 +46,59 @@ export default function SetupForm() {
 	const homeURL = useSelect( ( select ) => select( CORE_SITE ).getHomeURL() );
 
 	return (
-		<form>
-			<StoreErrorNotices
-				moduleSlug={ MODULES_SIGN_IN_WITH_GOOGLE }
-				storeName={ MODULES_SIGN_IN_WITH_GOOGLE }
-			/>
-			<p className="googlesitekit-margin-bottom-0">
-				{ __(
-					'Add your client ID here to complete setup:',
-					'google-site-kit'
-				) }
-			</p>
-			<div className="googlesitekit-setup-module__inputs">
-				<ClientIDTextField />
+		<div className="googlesitekit-sign-in-with-google-setup__form">
+			<div className="googlesitekit-setup-module__panel-item">
+				<StoreErrorNotices
+					moduleSlug={ MODULES_SIGN_IN_WITH_GOOGLE }
+					storeName={ MODULES_SIGN_IN_WITH_GOOGLE }
+				/>
+				<p className="googlesitekit-setup-module__step-description">
+					{ createInterpolateElement(
+						__(
+							'To set up Sign in With Google, Site Kit will help you create an "OAuth Client ID" that will be used to enable Sign in With Google on your website. You will be directed to a page that will allow you to generate an "OAuth Client ID". <a>Learn more</a>',
+							'google-site-kit'
+						),
+						{
+							a: <SupportLink path="#" external />,
+						}
+					) }
+				</p>
+				<p className="googlesitekit-margin-bottom-0">
+					{ __(
+						'Add your client ID here to complete setup:',
+						'google-site-kit'
+					) }
+				</p>
+				<div className="googlesitekit-setup-module__inputs">
+					<ClientIDTextField />
+				</div>
+				<Button
+					className="googlesitekit-sign-in-with-google-client-id-cta"
+					href={ `https://developers.google.com/web/site-kit?sitename=${ encodeURIComponent(
+						siteName
+					) }&siteurl=${ encodeURIComponent( homeURL ) }` }
+					target="_blank"
+					trailingIcon={ <ExternalIcon width="15" height="15" /> }
+					inverse
+				>
+					{ __( 'Get your client ID', 'google-site-kit' ) }
+				</Button>
 			</div>
-			<Button
-				className="googlesitekit-sign-in-with-google-client-id-cta"
-				href={ `https://developers.google.com/web/site-kit?sitename=${ encodeURIComponent(
-					siteName
-				) }&siteurl=${ encodeURIComponent( homeURL ) }` }
-				target="_blank"
-				trailingIcon={ <ExternalIcon width="15" height="15" /> }
-				inverse
-			>
-				{ __( 'Get your client ID', 'google-site-kit' ) }
-			</Button>
-		</form>
+
+			<div className="googlesitekit-setup-module__panel-item googlesitekit-setup-module__panel-item--with-svg">
+				<Suspense
+					fallback={ <PreviewBlock width="100%" height="235px" /> }
+				>
+					<MediaErrorHandler
+						errorMessage={ __(
+							'Failed to load graphic',
+							'google-site-kit'
+						) }
+					>
+						<LazyGraphicSVG />
+					</MediaErrorHandler>
+				</Suspense>
+			</div>
+		</div>
 	);
 }
